requirements.txt
asgiref==3.6.0
channels==4.0.0
channels-redis==4.0.0
Django==4.1.4
redis==4.4.0
sqlparse==0.4.3
uvicorn[standard]==0.20.0
djangorestframework-simplejwt==5.2.2
httpx==0.23.1
celery[redis]==5.2.7